requirements.txt
selenium==4.18.1
webdriver-manager==4.0.2
schedule==1.2.2
python-dotenv==1.0.1
pytest==8.3.3
pytest-cov==5.0.0
flake8==7.1.1
pyyaml==6.0.2
